fix(ci): restore dev route budgets (#1215)
